Updated controller styles to pass the linter
This included moving a large chunk of queries to a concern to decrease
the number of lines in control actions.

Updated the linter config to allow for methods up to 15 lines long.
Stubbed out some missing tests for controller actions that respond
differently to a letter parameter.
Fixed linting for ./config